Update TORSION_ATOMS_FRAGMENT after Omega call
Omega is used to sample secondary torsions for each of the torsion
fragments. However the omega call resets the atom indices on the
fragments making the original TORSION_ATOMS_FRAGMENT meaningless.

This change updates TORSION_ATOMS_FRAGMENT and TORSIN_ATOMPROP to point
to the new indices of the atoms of the primary torsion.